% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpm_cols.R
\name{cpm_cols}
\alias{cpm_cols}
\title{function to extract cpm colors as hex codes}
\usage{
cpm_cols(...)
}
\arguments{
\item{...}{Character names of cpm_colors}
}
\value{
The function returns the hex codes for the colors used in this package.
}
\description{
Extract the colors
}
\details{
Based on a blogpost on r-bloggers.com and the package ggtheme
}
\author{
Rico Derks
}
