% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcpm_data.R
\docType{data}
\name{ggcpm_data}
\alias{ggcpm_data}
\title{Palette data for ggCPM package}
\format{A \code{list}.}
\usage{
ggcpm_data
}
\description{
Data for colors and palettes
}
\details{
Data used by the palettes in the ggCPM package
}
\keyword{datasets}
