% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_cpm.R
\name{scale_fill_cpm}
\alias{scale_fill_cpm}
\title{Fill scale constructor for cpm colors}
\usage{
scale_fill_cpm(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in cpm palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_fill_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Set the fill palette for CPM.
}
\author{
Rico Derks
}
