% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cpm.R
\name{theme_cpm}
\alias{theme_cpm}
\title{Set the cpm theme for ggplot2}
\usage{
theme_cpm()
}
\description{
Set the cpm theme for ggplot2 at the moment this is only the minimal theme (theme_minimal)
}
\author{
Rico Derks
}
