% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cpm.R
\name{theme_cpm}
\alias{theme_cpm}
\title{Set the cpm theme for ggplot2}
\usage{
theme_cpm()
}
\description{
Set the cpm theme for ggplot2.
}
\details{
Set the CPM theme for ggplot2. The theme is based on theme_minimal()
}
\author{
Rico Derks
}
